Update adminplus code
