+++ fst_web/adminplus/templates/adminplus/index.html
@@ -9,7 +9,7 @@
       <caption>Lista föreskrifter och allmänna råd</caption>
       <tbody>
         {% for path, name in custom_list %}
-          <tr><td><a href="{{ path }}">{{ name }}</a></td></tr>
+           <tr><th><a href="{{ path }}">{{ name }}</a></th></tr>q
         {% endfor %}
       </tbody>
     </table>
