+++ requirements.txt
@@ -1,23 +1,4 @@
-astroid==1.4.6
-cffi==1.7.0
-colorama==0.3.7
-cryptography==1.4
 Django==1.9
-ecdsa==0.13
-Fabric==1.11.1
-html5lib==0.9999999
-idna==2.1
 isodate==0.5.4
-keepalive==0.5
-lazy-object-proxy==1.2.2
-logilab-common==1.2.2
-paramiko==2.0.1
-pyasn1==0.1.9
-pycparser==2.14
-pycrypto==2.6.1
-pylint==1.5.6
-pyparsing==2.1.5
+pyparsing==2.1.10
 rdflib==4.2.1
-six==1.10.0
-SPARQLWrapper==1.7.6
-wrapt==1.10.8
